requirements.txt
# 웹 프레임워크
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
httpx>=0.25.0

# 메시징 및 캐싱
redis>=5.0.0
pika>=1.3.2
aio-pika>=9.5.5

# LLM 통합
litellm>=1.12.0

# 유틸리티
python-dotenv>=1.0.0
loguru>=0.7.0

# 테스트
pytest>=7.0.0
pytest-asyncio>=0.20.0
